wip: logging improved
