+++ distage/distage-testkit-scalatest/src/main/scala/org/scalatest/distage/DistageScalatestTestSuiteRunner.scala
@@ -7,7 +7,7 @@ import izumi.distage.testkit.DebugProperties
 import izumi.distage.testkit.model.DistageTest
 import izumi.distage.testkit.runner.DistageTestRunner
 import izumi.distage.testkit.runner.api.TestReporter
-import izumi.distage.testkit.runner.services.{LateLoggerFactoryCachingImpl, ReporterBracket, TestkitLogging}
+import izumi.distage.testkit.runner.services.{ReporterBracket, TestkitLogging}
 import izumi.distage.testkit.services.scalatest.dstest.DistageTestsRegistrySingleton.SuiteReporter
 import izumi.distage.testkit.services.scalatest.dstest.{DistageTestsRegistrySingleton, SafeTestReporter}
 import izumi.distage.testkit.spec.AbstractDistageSpec
@@ -194,21 +194,20 @@ abstract class DistageScalatestTestSuiteRunner[F[_]](
     }
 
     // TODO: we may even share logger across F implementations, but that would be harder to do in a nice manner because there is no lifecycle in scalatest
-    val cache = LateLoggerFactoryCachingImpl.makeCache()
+//    val cache = LateLoggerFactoryCachingImpl.makeCache()
     try {
       if (toRun.nonEmpty) {
         debugLogger.log(s"GOING TO RUN TESTS in ${tagMonoIO.tag}: ${toRun.map(_.meta.id.name)}")
         val runner = new DistageTestRunner[F](
           testReporter,
           _.isInstanceOf[TestCanceledException],
-          cache,
           new ReporterBracket[F](testReporter),
           new TestkitLogging(),
         )
         runner.run(toRun)
       }
     } finally {
-      cache.close()
+//      cache.close()
       DistageTestsRegistrySingleton.completeStatuses[F]()
     }
   }
